[Fix] 에러뷰 뷰컨으로 변경 (#162)
